requirements.txt
## tensorflow
spyder
spyder-notebook
#tensorflow
tensorflow-cpu
numpy

## scikit-learn
#scikit-learn

##
#matplotlib
